# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gavin API",
    "description": "Content publishing and administration API for Gavin.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/assistant": {
      "get": {
        "tags": [
          "admin assistant"
        ],
        "summary": "Get Assistant Admin",
        "operationId": "get_assistant_admin_api_v1_admin_assistant_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantAdminSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assistant/sessions": {
      "post": {
        "tags": [
          "assistant"
        ],
        "summary": "Create Assistant Session",
        "operationId": "create_assistant_session_api_v1_assistant_sessions_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "assistant_session_missing|assistant_session_expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "origin_rejected|csrf_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict|event_cursor_invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "idempotency_result_expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited|concurrency_limited|budget_exhausted|stream_connection_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "assistant_disabled|assistant_not_ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assistant/session": {
      "get": {
        "tags": [
          "assistant"
        ],
        "summary": "Get Assistant Session",
        "operationId": "get_assistant_session_api_v1_assistant_session_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "assistant_session_missing|assistant_session_expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "origin_rejected|csrf_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict|event_cursor_invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "idempotency_result_expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited|concurrency_limited|budget_exhausted|stream_connection_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "assistant_disabled|assistant_not_ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "assistant"
        ],
        "summary": "Delete Assistant Session",
        "operationId": "delete_assistant_session_api_v1_assistant_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "assistant_session_missing|assistant_session_expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "origin_rejected|csrf_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "idempotency_conflict|event_cursor_invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "idempotency_result_expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "rate_limited|concurrency_limited|budget_exhausted|stream_connection_limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "assistant_disabled|assistant_not_ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "parameters": [
          {
            "name": "X-Assistant-CSRF",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/AdminErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "AdminErrorEnvelope"
      },
...
      "AssistantAdminSnapshot": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "partial"
            ],
            "title": "Status"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At"
          },
          "runtime_observed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Observed At"
          },
          "content_observed_at": {
            "type": "string",
            "title": "Content Observed At"
          },
          "deployment": {
            "$ref": "#/components/schemas/DeploymentView"
          },
          "availability": {
            "$ref": "#/components/schemas/AvailabilityView"
          },
          "readiness": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "cleanup": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "restore": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "qdrant": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "chat_provider": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "embedding_provider": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "worker": {
            "$ref": "#/components/schemas/ObservedFact"
          },
          "manifest": {
            "$ref": "#/components/schemas/ManifestView"
          },
          "queue": {
            "$ref": "#/components/schemas/QueueView"
          },
          "operation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "budgets": {
            "items": {
              "$ref": "#/components/schemas/BudgetView"
            },
            "type": "array",
            "title": "Budgets"
          },
          "daily_activity": {
            "$ref": "#/components/schemas/DailyActivityView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "observed_at",
          "runtime_observed_at",
          "content_observed_at",
          "deployment",
          "availability",
          "readiness",
          "cleanup",
          "restore",
          "qdrant",
          "chat_provider",
          "embedding_provider",
          "worker",
          "manifest",
          "queue",
          "operation",
          "budgets",
          "daily_activity"
        ],
        "title": "AssistantAdminSnapshot"
      },
...
      "AvailabilityView": {
        "properties": {
          "requested_state": {
            "type": "string",
            "enum": [
              "disabled",
              "enabled"
            ],
            "title": "Requested State"
          },
          "effective_state": {
            "type": "string",
            "enum": [
              "disabled",
              "enabled",
              "blocked"
            ],
            "title": "Effective State"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "operational_epoch": {
            "type": "integer",
            "title": "Operational Epoch"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "switch_pending_operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Switch Pending Operation Id"
          },
          "switch_target_generation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Switch Target Generation Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "draining_attempts": {
            "type": "integer",
            "title": "Draining Attempts"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requested_state",
          "effective_state",
          "version",
          "operational_epoch",
          "blocked_reason",
          "switch_pending_operation_id",
          "switch_target_generation_id",
          "updated_at",
          "draining_attempts"
        ],
        "title": "AvailabilityView"
      },
...
      "DeploymentView": {
        "properties": {
          "api_capability": {
            "type": "boolean",
            "title": "Api Capability"
          },
          "single_api_owner": {
            "type": "boolean",
            "title": "Single Api Owner"
          },
          "launcher_source": {
            "type": "string",
            "const": "web_runtime_config",
            "title": "Launcher Source"
          },
          "launcher_mounted": {
            "type": "null",
            "title": "Launcher Mounted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "api_capability",
          "single_api_owner",
          "launcher_source"
        ],
        "title": "DeploymentView"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "ManifestView": {
        "properties": {
          "generation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "dimension": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimension"
          },
          "pipeline_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Version"
          },
          "collection_present": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Present"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generation_id",
          "provider",
          "model",
          "model_version",
          "dimension",
          "pipeline_version",
          "collection_present"
        ],
        "title": "ManifestView"
      },
...
      "ObservedFact": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "healthy",
              "degraded",
              "blocked",
              "disabled",
              "stale",
              "unknown"
            ],
            "title": "Status"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "freshness_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freshness Seconds"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "reason_code",
          "observed_at",
          "freshness_seconds",
          "stale"
        ],
        "title": "ObservedFact"
      },
      "OperationView": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "kind": {
            "type": "string",
            "const": "rebuild",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "waiting",
              "catch_up",
              "ready_to_switch",
              "switch_pending",
              "switched",
              "failed",
              "abandoned"
            ],
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "generation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Id"
          },
          "previous_generation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Generation Id"
          },
          "source_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Cursor"
          },
          "outbox_high_water": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outbox High Water"
          },
          "safe_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "terminal_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "kind",
          "status",
          "version",
          "generation_id",
          "previous_generation_id",
          "source_cursor",
          "outbox_high_water",
          "safe_error_code",
          "message",
          "created_at",
          "updated_at",
          "terminal_at"
        ],
        "title": "OperationView"
      },
...
      "QueueView": {
        "properties": {
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "leased": {
            "type": "integer",
            "title": "Leased"
          },
          "succeeded": {
            "type": "integer",
            "title": "Succeeded"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "latest_success_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Success At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pending",
          "leased",
          "succeeded",
          "failed",
          "latest_success_at"
        ],
        "title": "QueueView"
      },
...
      "SessionCreateResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "idle_expires_at": {
            "type": "string",
            "title": "Idle Expires At"
          },
          "absolute_expires_at": {
            "type": "string",
            "title": "Absolute Expires At"
          },
          "policy": {
            "$ref": "#/components/schemas/AssistantClientPolicy"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "csrf_token",
          "idle_expires_at",
          "absolute_expires_at",
          "policy"
        ],
        "title": "SessionCreateResponse"
      },
...
      "SessionStatusResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "idle_expires_at": {
            "type": "string",
            "title": "Idle Expires At"
          },
          "absolute_expires_at": {
            "type": "string",
            "title": "Absolute Expires At"
          },
          "policy": {
            "$ref": "#/components/schemas/AssistantClientPolicy"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/SessionTurn"
            },
            "type": "array",
            "title": "Turns"
          },
          "active_turn": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActiveTurn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "created_at",
          "idle_expires_at",
          "absolute_expires_at",
          "policy",
          "turns"
        ],
        "title": "SessionStatusResponse"
      },
...
    }
  }
}